requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
build==1.2.2.post1
colorama==0.4.6
crispy-bootstrap5==2025.4
Django==5.0.6
django-crispy-forms==2.4
django-widget-tweaks==1.5.0
djangorestframework==3.16.0
greenlet==3.0.1
iniconfig==2.1.0
numpy==1.26.0
packaging==25.0
playwright==1.40.0
playwright-stealth==1.0.6
pluggy==1.6.0
pydantic==2.5.3
pydantic_core==2.14.6
pyee==11.0.1
pyproject_hooks==1.2.0
pytest==7.4.0
pytest-asyncio==0.23.5
python-dotenv==1.0.0
setuptools==80.9.0
sqlparse==0.5.0
typing_extensions==4.13.2
tzdata==2024.1
websockets==12.0
wheel==0.45.1
